+++ src/main/java/com/leokom/games/chess/player/legal/brain/common/GenericEvaluator.java
@@ -1,6 +1,6 @@
 package com.leokom.games.chess.player.legal.brain.common;
 
-import com.leokom.games.chess.engine.GameState;
+import com.leokom.games.commons.engine.GameState;
 import com.leokom.games.commons.engine.GameTransition;
 
 /**
